Update UIImage+Gif.swift
- Changed the type of index parameter of delayForImageAtIndex() to Int from UInt, since CGImageSourceCopyPropertiesAtIndex() accepts index as Int.
- Removed the UInt cast at the line 111.

+++ SwiftGifCommon/UIImage+Gif.swift
@@ -18,7 +18,7 @@ extension UIImage {
         return image
     }
     
-    class func delayForImageAtIndex(index: UInt, source: CGImageSourceRef)
+    class func delayForImageAtIndex(index: Int, source: CGImageSourceRef)
         -> Double {
         var delay = 0.1
         
@@ -108,7 +108,7 @@ extension UIImage {
             images.append(CGImageSourceCreateImageAtIndex(source, i, nil))
             
             // At it's delay in cs
-            var delaySeconds = UIImage.delayForImageAtIndex(UInt(i),
+            var delaySeconds = UIImage.delayForImageAtIndex(i,
                 source: source)
             delays.append(Int(delaySeconds * 1000.0)) // Seconds to ms
         }
